requirements.txt
opencv-python==4.10.0.84
torch==2.0.1+cu117
pandas==1.5.3
openpyxl==3.1.2
transformers==4.30.0
matplotlib==3.5.1
numpy==1.24.3
python-Levenshtein